fix: standardise UUID defaults and shared_org_id type in migration
- Replace gen_random_uuid() with uuid_generate_v4() on user_inventory_item
  and org_inventory_item — gen_random_uuid() requires pgcrypto which is
  not guaranteed on a clean Postgres instance; uuid_generate_v4() is
  provided by the already-enabled uuid-ossp extension
- Change UserInventoryItem.sharedOrgId column type from bigint to int to
  match Organization.id (SERIAL = INTEGER); bigint created a type mismatch
  across the FK that could cause serialization surprises

+++ backend/src/modules/user-inventory/entities/user-inventory-item.entity.ts
@@ -82,7 +82,7 @@ export class UserInventoryItem {
   @Column({ type: 'text', nullable: true })
   notes?: string;
 
-  @Column({ name: 'shared_org_id', type: 'bigint', nullable: true })
+  @Column({ name: 'shared_org_id', type: 'int', nullable: true })
   @Index()
   sharedOrgId?: number | null;
